fix(RT-2): Round 2 corrections — all 7 blockers addressed
1. GCC sign-compare: cast int literal to size_t in strings.cpp
2. Single SHA-256 replay context: removed full_buf and second replay
3. Directory discovery: report all .mxraw/.mxraw.partial candidates
4. Session selection: replay_from_stream_set() for explicit selection
5. 64-bit file position: fseek64/ftell64 for 64 GiB limit
6. Short-read checks: IoError on premature EOF in SHA loops
7. New tests: resource boundedness, discovery, two-session, ambiguity

RT-2: 15/15 Windows Release
RT-1: 6/6 (no regression)
QSH/M10X: 20/20 (no regression)
Python: 3/3
Hygiene: PASS

+++ cpp/moex_raw/src/strings.cpp
@@ -68,10 +68,10 @@ bool read_length_string(const std::uint8_t* data, std::size_t available,
 
     std::uint16_t len = read_u16_le(data);
     if (len > kMaxStringBytes) return false;
-    if (available < 2 + len) return false;
+    if (available < static_cast<std::size_t>(2) + len) return false;
 
     out.assign(reinterpret_cast<const char*>(data + 2), len);
-    bytes_consumed = 2 + len;
+    bytes_consumed = static_cast<std::size_t>(2) + len;
 
     // Validate UTF-8
     if (!validate_utf8_string(out)) return false;
